+++ ui/client/analytics/AnalyticsEngine.js
@@ -1,0 +1,5 @@
+export default class AnalyticsEngine {
+  sendEvent = (event) => {
+    throw new Error('You have to implement method send event!')
+  }
+}
