+++ ui/client/containers/Visualization.js
@@ -14,7 +14,7 @@ import SpinnerWrapper from "../components/SpinnerWrapper";
 import * as JsonUtils from "../common/JsonUtils";
 import RouteLeavingGuard from "../components/RouteLeavingGuard";
 import ClipboardUtils from "../common/ClipboardUtils";
-import {events} from "../analytics/matomo/TrackingEvents"
+import {events} from "../analytics/TrackingEvents"
 
 class Visualization extends React.Component {
